Adding InstallUtilities and Add-ScoopBucket

+++ bootstrapScripts/02-setupForInstalls.ps1
@@ -4,10 +4,17 @@
 Set-StrictMode -Version 2.0
 
 # Copy paste.exe to C:\Windows\system32 directory
-Copy-Item -Path "$env:ProfilePath\winfiles\installs\paste.exe" -Destination "$env:SystemRoot\system32"
+if (-not (Test-Path "$env:SystemRoot\system32\paste.exe")) {
+    Write-Host "Copying paste.exe"
+    Copy-Item -Path "$env:ProfilePath\winfiles\installs\paste.exe" -Destination "$env:SystemRoot\system32"
+}
 
 # Setup "extras" repo for scoop
-Invoke-Expression "scoop bucket add extras"
+Write-Host "Setting up scoop buckets"
+Add-ScoopBucket "main"
+Add-ScoopBucket "extras"
+Add-ScoopBucket "nerd-fonts"
+Add-ScoopBucket "java"
 
 # Install AppGet
 #Invoke-Expression "$env:ProfilePath\winfiles\installs\appget.exe"
